+++ joulescope_ui/error_window.py
@@ -17,6 +17,7 @@ from joulescope_ui import reporter
 from joulescope_ui import versioned_file
 from joulescope_ui.zip_inspector import ZipInspectorDialog
 from PySide6 import QtCore, QtGui, QtWidgets
+import datetime
 import json
 import markdown
 import os
@@ -31,16 +32,34 @@ _CONTACT_TEXT = N_("""\
 _HTML = "<html><head></head><body>{body}</body></html>"
 _RECOVERY = N_("Select an error recovery option.")
 _CONTACT_FILE = os.path.join(pubsub_singleton.query('common/settings/paths/config'), 'contact.json')
+_RESPONSE_TIME_DAYS = 2
+_TYPICAL_RESPONSE_TIME = N_(f'Typical response time: {_RESPONSE_TIME_DAYS} business days')
+_HOLIDAY = N_("""\
+Our offices are closed for holiday from {start_date} to {end_date}.
+We will respond within {response_time} business days when we return on {return_date}.""")
+_HOLIDAYS = {
+    # (datetime.date.today(), datetime.date.today() + datetime.timedelta(7)),  # for testing
+    (datetime.date(2023, 8, 12), datetime.date(2023, 8, 27)),
+    (datetime.date(2023, 9, 2), datetime.date(2023, 9, 4)),
+    (datetime.date(2023, 11, 18), datetime.date(2023, 11, 26)),
+    (datetime.date(2023, 12, 22), datetime.date(2024, 1, 1)),
+    (datetime.date(2024, 2, 17), datetime.date(2024, 2, 19)),
+}
 
 
 class SubmitThread(QtCore.QThread):
 
+    def __init__(self):
+        self.results = None
+        super().__init__()
+
     def run(self):
-        reporter.publish()
+        self.results = reporter.publish()
 
 
-class SubmitWidget(QtWidgets.QWidget):
+class SubmitConfigureWidget(QtWidgets.QWidget):
 
+    started = QtCore.Signal()
     finished = QtCore.Signal()
 
     def __init__(self, parent, report_path):
@@ -87,7 +106,7 @@ class SubmitWidget(QtWidgets.QWidget):
         self._description_view.setWordWrap(True)
         self._description_view.setAlignment(QtCore.Qt.AlignTop)
         self._description_tabs.addTab(self._description_edit, N_('Edit'))
-        self._description_tabs.addTab(self._description_view, N_('View'))
+        self._description_tabs.addTab(self._description_view, N_('View as Markdown'))
         self._description_layout.addWidget(self._description_tabs)
         self._description.setLayout(self._description_layout)
         self._layout.addWidget(self._description)
@@ -109,6 +128,15 @@ class SubmitWidget(QtWidgets.QWidget):
         self._submit.pressed.connect(self._on_submit)
 
         self.setLayout(self._layout)
+
+    @property
+    def results(self):
+        r = None
+        if self._thread is not None:
+            r = self._thread.results
+        if r is None:
+            return ''
+        return '\n'.join(r)
 
     def _on_description_tab_changed(self, index):
         if index == 1:
@@ -138,12 +166,94 @@ class SubmitWidget(QtWidgets.QWidget):
             json.dump(contact, f)
         description = self._description_edit.toPlainText()
         reporter.update_description(self._report_path, description)
+        self.started.emit()
         self._thread = SubmitThread()
         self._thread.finished.connect(self._on_submit_finished)
         self._thread.start()
 
     @QtCore.Slot()
     def _on_submit_finished(self):
+        self.finished.emit()
+
+
+class SubmitStatusWidget(QtWidgets.QWidget):
+
+    finished = QtCore.Signal()
+
+    def __init__(self, parent):
+        super().__init__(parent=parent)
+        self._layout = QtWidgets.QVBoxLayout(self)
+        self._status_label = QtWidgets.QLabel(N_('Submit in progress.'), self)
+        self._results_label = QtWidgets.QLabel(self)
+        self._results_label.setWordWrap(True)
+        self._response_label = QtWidgets.QLabel(self)
+        self._response_label.setWordWrap(True)
+
+        self._layout.addWidget(self._status_label)
+        self._layout.addWidget(self._results_label)
+        self._layout.addWidget(self._response_label)
+
+        self._spacer = QtWidgets.QSpacerItem(0, 0, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self._layout.addItem(self._spacer)
+
+        self._buttons = None
+
+    def results(self, results):
+        self._status_label.setText(N_('Submit completed.'))
+        self._results_label.setText(results)
+        self._update_response()
+        self._buttons = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok)
+        self._buttons.accepted.connect(self.finished.emit)
+        self._buttons.rejected.connect(self.finished.emit)
+        self._layout.addWidget(self._buttons)
+
+    def _update_response(self):
+        today = datetime.date.today()
+        today_plus = today + datetime.timedelta(_RESPONSE_TIME_DAYS)
+        for d_start, d_end in _HOLIDAYS:
+            if today_plus >= d_start and today <= d_end:
+                txt = _HOLIDAY.format(
+                    start_date=d_start.isoformat(),
+                    end_date=d_end.isoformat(),
+                    return_date=(d_end + datetime.timedelta(1)).isoformat(),
+                    response_time=_RESPONSE_TIME_DAYS,
+                )
+                self._response_label.setText(txt)
+                return
+        self._response_label.setText(_TYPICAL_RESPONSE_TIME)
+
+
+class SubmitWidget(QtWidgets.QWidget):
+
+    finished = QtCore.Signal()
+
+    def __init__(self, parent, report_path):
+        self._parent = parent
+        self._thread = None
+        super().__init__(parent=parent)
+        self._layout = QtWidgets.QVBoxLayout(self)
+
+        self._config = SubmitConfigureWidget(parent, report_path)
+        self._config.started.connect(self._on_submit_started)
+        self._config.finished.connect(self._on_config_finished)
+        self._layout.addWidget(self._config)
+
+        self._status = SubmitStatusWidget(parent)
+        self._status.finished.connect(self._on_status_finished)
+        self._status.hide()
+
+    def _on_submit_started(self):
+        self._config.hide()
+        self._layout.removeWidget(self._config)
+        self._layout.addWidget(self._status)
+        self._status.show()
+
+    def _on_config_finished(self):
+        self._status.results(self._config.results)
+
+    def _on_status_finished(self):
+        self._status.hide()
+        self._layout.removeWidget(self._status)
         self.finished.emit()